+++ lib/eva/sfa/sfapi.py
@@ -622,10 +622,6 @@ class SFA_HTTP_API(GenericHTTP_API, SFA_API):
         return result
 
     def action(self, k=None, i=None, u=None, s=None, v='', p=None, q=None, w=0):
-        try:
-            _s = int(s)
-        except:
-            raise cp_api_error('status is not an integer')
         if w:
             try:
                 _w = float(w)
@@ -651,7 +647,7 @@ class SFA_HTTP_API(GenericHTTP_API, SFA_API):
             k=k,
             i=i,
             action_uuid=u,
-            nstatus=_s,
+            nstatus=s,
             nvalue=v,
             priority=p,
             q=q,