+++ lib/eva/uc/controller.py
@@ -572,9 +572,15 @@ def exec_unit_action(unit,
     else:
         u = unit
     if not u: return None
+    _s = None
+    try:
+        _s = int(nstatus)
+    except:
+        _s = u.status_by_label(nstatus)
+    if _s is None: return None
     if q_timeout: qt = q_timeout
     else: qt = eva.core.timeout
-    a = u.create_action(nstatus, nvalue, priority, action_uuid)
+    a = u.create_action(_s, nvalue, priority, action_uuid)
     Q.put_task(a)
     if not eva.core.wait_for(a.is_processed, q_timeout):
         if a.set_dead():