+++ CHANGELOG.md
@@ -14,6 +14,8 @@ UC:
     * new function state_history in UC API
     * UC EI now should be enabled/disabled in uc.ini
     * EVA_ITEM_OID var in the environment of UC scripts
+    * action status label (case insensitive) may be used instead of number.
+      if the label is not defined, API return 404 error
 
 LM: